Splited model into two
