+++ NeuroTradeAPI/NeuroTradeAPI/Program.cs
@@ -15,6 +15,7 @@ namespace NeuroTradeAPI
         public static void Main(string[] args)
         {
             BuildWebHost(args).Run();
+            
         }
 
         public static IWebHost BuildWebHost(string[] args) =>
